third_party/blink/web_tests/platform/win/external/wpt/service-workers/service-worker/navigation-timing-extended.https-expected.txt

This is a testharness.js-based test.
FAIL Service worker controlled navigation timing assert_greater_than: workerStart marking should not wait for worker activation to finish expected a number greater than 1634044893684.5999 but got 1634044893683.9001
Harness: the test ran to completion.
